springbook/src/test/java/springbook/user/dao/UserDaoTest.java
package springbook.user.dao;

import org.junit.Test;
import org.junit.runner.JUnitCore;
import org.springframework.boot.test.util.ApplicationContextTestUtils;
import org.springframework.context.ApplicationContext;
import org.springframework.context.annotation.AnnotationConfigApplicationContext;
import springbook.user.domain.User;

import java.sql.SQLException;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

/**
 * Created by geekslife on 2017. 1. 8..
 */
public class UserDaoTest {
    @Test
    public void addAndGet() throws SQLException, ClassNotFoundException {
        ApplicationContext context = new AnnotationConfigApplicationContext(DaoFactory.class);
        UserDao dao = context.getBean("userDao",UserDao.class);

        User user = new User();
        user.setId("gyumee");
        user.setName("박성철");
        user.setPassword("springno1");

        dao.add( user );
        assertThat(dao.getCount(),is(1));

        User user2 = dao.get(user.getId());

        assertThat(user2.getName(), is(user.getName()));
        dao.deleteAll();
        assertThat(dao.getCount(),is(0));
    }

    @Test
    public void count() throws SQLException, ClassNotFoundException {
        ApplicationContext context = new AnnotationConfigApplicationContext(DaoFactory.class);
        UserDao dao = context.getBean("userDao",UserDao.class);

        assertThat(dao.getCount(), is(0));

        User user = new User("hyumee","박성철", "aa");
        user.setId("gyumee");
        user.setName("박성철");
        user.setPassword("springno1");
        dao.add(user);

        assertThat(dao.getCount(), is(1));


        User user2 = new User();
        user2.setId("geekslife");
        user2.setName("ㅣ  ");
        user2.setPassword("xxx");
        dao.add(user2);

        assertThat(dao.getCount(), is(2));

        dao.deleteAll();
    }

    public static void main(String[] args) throws SQLException, ClassNotFoundException {
        JUnitCore.main("springbook.user.dao.UserDaoTest");
    }

}